refactor(fast-element): relocate export modules into feature directories
