--- packages/fast-element/src/binding.ts
@@ -1,7 +1,0 @@
-export { Binding, type BindingDirective } from "./binding/binding.js";
-export { normalizeBinding } from "./binding/normalize.js";
-export { oneTime } from "./binding/one-time.js";
-export { listener, oneWay } from "./binding/one-way.js";
-export type { DOMAspect, DOMPolicy } from "./dom.js";
-export type { Subscriber } from "./observation/notifier.js";
-export type { Expression, ExpressionObserver } from "./observation/observable.js";
