--- packages/fast-element/src/css.ts
@@ -1,10 +1,0 @@
-export { type CSSTemplateTag, type CSSValue, css } from "./styles/css.js";
-export {
-    CSSDirective,
-    type CSSDirectiveDefinition,
-    cssDirective,
-} from "./styles/css-directive.js";
-export {
-    type ComposableStyles,
-    ElementStyles,
-} from "./styles/element-styles.js";
