--- packages/fast-element/src/directives/repeat.ts
@@ -1,9 +1,0 @@
-export type { Binding } from "../binding/binding.js";
-export type { Expression } from "../observation/observable.js";
-export {
-    RepeatBehavior,
-    RepeatDirective,
-    type RepeatOptions,
-    repeat,
-} from "../templating/repeat.js";
-export type { CaptureType, ViewTemplate } from "../templating/template.js";
