--- packages/fast-element/src/directives/when.ts
@@ -1,3 +1,0 @@
-export type { ExecutionContext, Expression } from "../observation/observable.js";
-export type { CaptureType, SyntheticViewTemplate } from "../templating/template.js";
-export { when } from "../templating/when.js";
