--- packages/fast-element/src/html.ts
@@ -1,14 +1,0 @@
-export type { Binding } from "./binding/binding.js";
-export type { DOMPolicy } from "./dom.js";
-export type { Expression } from "./observation/observable.js";
-export {
-    type ElementViewTemplate,
-    type HTMLTemplateCompilationResult,
-    type HTMLTemplateTag,
-    html,
-    InlineTemplateDirective,
-    type SyntheticViewTemplate,
-    type TemplateValue,
-    ViewTemplate,
-} from "./templating/template.js";
-export { type ElementView, HTMLView, type SyntheticView } from "./templating/view.js";
