--- packages/fast-element/src/hydration.ts
@@ -1,71 +1,0 @@
-export type { HydrationOptions } from "./components/enable-hydration.js";
-export {
-    deferHydrationAttribute,
-    enableHydration,
-    StopHydration,
-} from "./components/enable-hydration.js";
-export { isHydratable } from "./components/hydration.js";
-export { HydrationTracker } from "./components/hydration-tracker.js";
-export type {
-    DOMAspect,
-    DOMSink,
-} from "./dom.js";
-export {
-    type DOMAspectGuards,
-    type DOMElementGuards,
-    type DOMGuards,
-    DOMPolicy,
-    type DOMPolicyOptions,
-    type DOMSinkGuards,
-} from "./dom-policy.js";
-export type {
-    HydrationDiagnostic,
-    HydrationDiagnosticResult,
-    HydrationMismatchActual,
-    HydrationMismatchExpectation,
-} from "./hydration/diagnostics.js";
-export type { HydrationDebugger } from "./hydration/hydration-debugger.js";
-export { hydrationDebugger } from "./hydration/hydration-debugger.js";
-export type {
-    ViewBehaviorBoundaries,
-    ViewBoundaries,
-} from "./hydration/target-builder.js";
-export type { Disposable, TrustedTypesPolicy } from "./interfaces.js";
-export type {
-    ExecutionContext,
-    ExpressionController,
-    SourceLifetime,
-} from "./observation/observable.js";
-export type {
-    ContentTemplate,
-    ContentView,
-    HydratableContentTemplate,
-} from "./templating/html-binding-directive.js";
-export type {
-    ViewBehavior,
-    ViewBehaviorFactory,
-    ViewBehaviorTargets,
-    ViewController,
-} from "./templating/html-directive.js";
-export type {
-    HydratableView,
-    HydratableViewController,
-    ViewNodes,
-} from "./templating/hydration-view.js";
-export {
-    HydrationBindingError,
-    HydrationStage,
-} from "./templating/hydration-view.js";
-export type {
-    CaptureType,
-    ElementViewTemplate,
-    HydratableElementViewTemplate,
-    HydratableSyntheticViewTemplate,
-    SyntheticViewTemplate,
-} from "./templating/template.js";
-export type {
-    DefaultExecutionContext,
-    ElementView,
-    SyntheticView,
-    View,
-} from "./templating/view.js";
