--- packages/fast-element/src/observable.ts
@@ -1,16 +1,0 @@
-export type { Notifier, Subscriber } from "./observation/notifier.js";
-export { PropertyChangeNotifier, SubscriberSet } from "./observation/notifier.js";
-export type {
-    Accessor,
-    Expression,
-    ExpressionController,
-    ExpressionNotifier,
-    ExpressionObserver,
-    ObservationRecord,
-} from "./observation/observable.js";
-export {
-    ExecutionContext,
-    Observable,
-    observable,
-    SourceLifetime,
-} from "./observation/observable.js";
