--- packages/fast-element/src/styles.ts
@@ -1,18 +1,0 @@
-export type { Constructable } from "./interfaces.js";
-export type {
-    ExecutionContext,
-    ExpressionController,
-    SourceLifetime,
-} from "./observation/observable.js";
-export {
-    CSSDirective,
-    type CSSDirectiveDefinition,
-    cssDirective,
-} from "./styles/css-directive.js";
-export {
-    type ComposableStyles,
-    type ConstructibleStyleStrategy,
-    ElementStyles,
-} from "./styles/element-styles.js";
-export type { HostBehavior, HostController } from "./styles/host.js";
-export type { StyleStrategy, StyleTarget } from "./styles/style-strategy.js";
