--- packages/fast-element/src/templating.ts
@@ -1,45 +1,0 @@
-export type { Binding, BindingDirective } from "./binding/binding.js";
-export type { DOMAspect, DOMPolicy } from "./dom.js";
-export type {
-    ExecutionContext,
-    ExpressionController,
-    SourceLifetime,
-} from "./observation/observable.js";
-export { type CompilationStrategy, Compiler } from "./templating/compiler.js";
-export {
-    type ContentTemplate,
-    type ContentView,
-    HTMLBindingDirective,
-} from "./templating/html-binding-directive.js";
-export {
-    type AddViewBehaviorFactory,
-    type Aspected,
-    type CompiledViewBehaviorFactory,
-    HTMLDirective,
-    type HTMLDirectiveDefinition,
-    htmlDirective,
-    type PartialHTMLDirectiveDefinition,
-    StatelessAttachedAttributeDirective,
-    type ViewBehavior,
-    type ViewBehaviorFactory,
-    type ViewBehaviorTargets,
-    type ViewController,
-} from "./templating/html-directive.js";
-export { Markup, Parser } from "./templating/markup.js";
-export { RenderBehavior, RenderDirective, render } from "./templating/render.js";
-export {
-    type CaptureType,
-    type ElementViewTemplate,
-    type HTMLTemplateCompilationResult,
-    type HTMLTemplateTag,
-    InlineTemplateDirective,
-    type SyntheticViewTemplate,
-    type TemplateValue,
-    ViewTemplate,
-} from "./templating/template.js";
-export {
-    type ElementView,
-    HTMLView,
-    type SyntheticView,
-    type View,
-} from "./templating/view.js";
