+++ packages/fast-element/src/templating/html.ts
@@ -1,0 +1,1 @@
+export { html } from "./template.js";
