--- packages/fast-element/src/volatile.ts
@@ -1,2 +1,0 @@
-export type { Accessor } from "./observation/observable.js";
-export { volatile } from "./observation/observable.js";
